Fixed a lot of cypress tests and actual build, not far now
