+++ src/features/CourtCaseDetails/CourtCaseDetails.stories.tsx
@@ -1,4 +1,5 @@
 import parseAhoXml from "@moj-bichard7-developers/bichard7-next-core/build/src/parse/parseAhoXml/parseAhoXml"
+import { AnnotatedHearingOutcome } from "@moj-bichard7-developers/bichard7-next-core/build/src/types/AnnotatedHearingOutcome"
 import { expect } from "@storybook/jest"
 import { ComponentMeta, ComponentStory } from "@storybook/react"
 import { within } from "@storybook/testing-library"
@@ -24,23 +25,24 @@ const courtCase = {
   ptiurn: "42CY0300107",
   triggerReason: "TRPR0006",
   triggers: [{ triggerCode: "TRPR0001" } as unknown as Trigger],
-  courtDate: new Date("2008-09-26"),
-  hearingOutcome: parseAhoXml(CourtCaseAho.annotated_msg)
+  courtDate: new Date("2008-09-26")
 } as unknown as CourtCase
 
-export const Details: ComponentStory<typeof CourtCaseDetails> = () => <CourtCaseDetails courtCase={courtCase} />
+const aho = parseAhoXml(CourtCaseAho.hearingOutcomeXml) as AnnotatedHearingOutcome
+
+export const Details: ComponentStory<typeof CourtCaseDetails> = () => (
+  <CourtCaseDetails courtCase={courtCase} aho={aho} />
+)
 
 Details.play = async ({ canvasElement }) => {
   const canvas = within(canvasElement)
-  await expect(canvas.getByText(courtCase.ptiurn)).toBeInTheDocument()
-  await expect(canvas.getByText(courtCase.courtName)).toBeInTheDocument()
+  await expect(canvas.getByText(courtCase.ptiurn!)).toBeInTheDocument()
+  await expect(canvas.getByText(courtCase.courtName!)).toBeInTheDocument()
   // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
   await expect(canvas.getByText(format(courtCase.courtDate!, "dd/MM/yyyy"))).toBeInTheDocument()
-  await expect(canvas.getByText(courtCase.defendantName)).toBeInTheDocument()
-  await expect(canvas.getByText(courtCase.triggerReason)).toBeInTheDocument()
+  await expect(canvas.getByText(courtCase.defendantName!)).toBeInTheDocument()
+  await expect(canvas.getByText(courtCase.triggerReason!)).toBeInTheDocument()
   await expect(
-    canvas.getByText(
-      courtCase.hearingOutcome.AnnotatedHearingOutcome.HearingOutcome.Hearing.CourtHearingLocation.OrganisationUnitCode
-    )
+    canvas.getByText(aho.AnnotatedHearingOutcome.HearingOutcome.Hearing.CourtHearingLocation.OrganisationUnitCode)
   ).toBeInTheDocument()
 }
