--- src/services/entities/transformers/booleanTransformer.ts
@@ -1,8 +1,0 @@
-import { ValueTransformer } from "typeorm"
-
-const booleanTransformer: ValueTransformer = {
-  to: (value) => (value ? 1 : 0),
-  from: (value) => value === 1
-}
-
-export default booleanTransformer
