+++ test/testFixtures/database/insertCourtCases.ts
@@ -2,47 +2,7 @@ import CourtCase from "../../../src/services/entities/CourtCase"
 import { InsertResult } from "typeorm"
 import getDataSource from "../../../src/services/getDataSource"
 import DummyAho from "./data/error_list_aho.json"
-
-const DummyCourtCase: Partial<CourtCase> = {
-  courtName: "Magistrates' Courts Essex Basildon",
-  courtDate: new Date("2008-09-26"),
-  ptiurn: "42CY0300108",
-  defendantName: "Allocation Trigger",
-  errorReason: "",
-  triggerReason: "",
-  errorId: 79507,
-  messageId: "cf60600b-52a0-4bc7-b117-3acfb851a7f5",
-  phase: 2,
-  errorStatus: 2,
-  triggerStatus: 1,
-  errorQualityChecked: 1,
-  triggerQualityChecked: 1,
-  triggerCount: 0,
-  errorLockedById: null,
-  triggerLockedById: null,
-  urgency: 1,
-  asn: "0836FP0100000377244A",
-  courtCode: "B42AZ01",
-  updatedHearingOutcome: null,
-  errorReport: "",
-  createdTimestamp: new Date("2022-06-30 09:44:03.93"),
-  errorCount: 0,
-  userUpdatedFlag: true,
-  resolutionTimestamp: null,
-  messageReceivedTimestamp: new Date("2022-06-30 09:44:03.93"),
-  errorResolvedBy: "essex.user",
-  triggerResolvedBy: null,
-  errorResolvedTimestamp: new Date("2022-06-30 09:44:11.182"),
-  triggerResolvedTimestamp: null,
-  orgForPoliceFilter: "36FP  ",
-  courtRoom: "01",
-  courtReference: "42CY0300108",
-  errorInsertedTimestamp: new Date("2022-06-30 09:44:03.99"),
-  triggerInsertedTimestamp: new Date("2022-06-30 09:44:11.193"),
-  pncUpdateEnabled: "Y",
-  notes: [],
-  triggers: []
-}
+import DummyCourtCase from "./DummyCourtCase"
 
 const getDummyCourtCase = async (overrides?: Partial<CourtCase>): Promise<CourtCase> => {
   const dataSource = await getDataSource()
